+++ NEWS.md
@@ -1,5 +1,10 @@
 # xpose.xtras (development version)
 
+# xpose.xtras 0.1.0
+
+* Bugs regarding covariate plotting functions were resolved.
+* nlmixr2 compatibility implemented
+
 # xpose.xtras 0.0.2
 
 * CRAN release planned